automaticResponseUrl can be different then normalReturnUrl
I'm not sure what the best way is. The automaticResponseUrl can be different then the normalReturnUrl. So I guess the automatic one is supposed to be the NotifyUrl? But it would be more developer-friendly to fallback on the ReturnUrl by default.
Or should is also be possible to disable the automatic response?